mysql: Keep reading bookmarks from primary in bookmarks and hooks code
Summary:
## This stack
In D81238781 I changed the `get` method in the `Bookmarks` facet to take a `Freshness` arg.
This was needed to then change the behaviour of always reading from the primary in D81212709.

I wanted to make the change no-op, so I updated all the callsites adding `Freshness::Recent`, which I would then later re-evaluate.

The problem is that there are **a lot of callsites**. Too many to evaluate one-by-one.
Besides that, most of the problematic calls are coming through SCS, via `repo_resolve_bookmark` or by calls to `mononoke_api::resolve_bookmark`, which are already updated to propagate the freshness.

In this stack I evaluated parts of our codebase **in bulk**, then I either changed the Freshness arg or removed the TODO comment.

## This diff
I'm considering all the code under `eden/mononoke/bookmarks` to be in a lower level of abstraction, so leaving it with MostRecent.

I also added pushrebase, slow_bookmark_mover and hooks code.

markbt please LMK if there's anything here that you think is high-traffic enough and safe to change to MaybeStale and I can remove it from this diff and evaluate it separately.

Differential Revision: D81447233

fbshipit-source-id: 02ec3a6c4b0544448b8e4c0f427c12f8958dfbc7