Test case to demo inferred_copy_from derivation issue
Summary:
I noticed this while investigating another blame issue.
This particular one doesn't actually affect the final outcome (it's pointing to itself in parent), but it added useless indirections when we consume the derived data.

This diff adds the test case to demo the issue. Fix is in next one.

Reviewed By: RajivTS

Differential Revision: D81531370

fbshipit-source-id: b210e8ab21f38ade22d73711f778a8d623e479a7