Fix inferred_copy_from derivation
Summary:
I noticed this while investigating another blame issue.
This particular one doesn't actually affect the final outcome (it's pointing to itself in parent), but it added useless indirections when we consume the derived data.

The issue boils down to we are comparing a file to its previous self during partial matching (as we grab the candidates in the previous round via multiple heuristics). The fix is simply filtering these paths out.

Reviewed By: RajivTS

Differential Revision: D81537131

fbshipit-source-id: 0e41bf792a54d12928d21f8f87bf52126f3c88e9