Move SQL storage helper library to common
Summary:
This started as an internal detail of `git_server/push_poller` but is actually a generic way to create a SqlConnection, for instance to `xdb.mononoke_production` without having to load the `Repo` trait.

Its usage already has spread to `mononoke_admin` and multiple `metagit` crates on top of the `push_poller`.

Make it common.

Also move `ephemeral_storage` which was its own crate and a bit of an implementation detail of `sql_storage` to a `mod` in `sql_storage`.

This will also make more sense when we start using it from SCS for automatic repo creation in Mononoke.

Reviewed By: RajivTS

Differential Revision: D81488418

fbshipit-source-id: 8c951a57d8ac2e7e51bf75045e2bd2fdb3cbd74f