mysql: Keep reading bookmarks from primary in blobrepo
Summary:
## This stack
In D81238781 I changed the `get` method in the `Bookmarks` facet to take a `Freshness` arg.
This was needed to then change the behaviour of always reading from the primary in D81212709.

I wanted to make the change no-op, so I updated all the callsites adding `Freshness::Recent`, which I would then later re-evaluate.

The problem is that there are **a lot of callsites**. Too many to evaluate one-by-one.
Besides that, most of the problematic calls are coming through SCS, via `repo_resolve_bookmark` or by calls to `mononoke_api::resolve_bookmark`, which are already updated to propagate the freshness.

In this stack I evaluated parts of our codebase **in bulk**, then I either changed the Freshness arg or removed the TODO comment.

## This diff

## RFC: blobrepo
This seems to be old code and risky to change, as I'm not aware of how/where it's used. I prefer to err on the side of caution and leave it unchanged, but would appreciate input from markbt

Differential Revision: D81582876

fbshipit-source-id: 3577d2abb5d67aeb808f6feb2f14b3e7c94d95af