mysql: Get bookmark value from replicas in walker and repo_stats_logger
Summary:
## This stack
In D81238781 I changed the `get` method in the `Bookmarks` facet to take a `Freshness` arg.
This was needed to then change the behaviour of always reading from the primary in D81212709.

I wanted to make the change no-op, so I updated all the callsites adding `Freshness::Recent`, which I would then later re-evaluate.

The problem is that there are **a lot of callsites**. Too many to evaluate one-by-one.
Besides that, most of the problematic calls are coming through SCS, via `repo_resolve_bookmark` or by calls to `mononoke_api::resolve_bookmark`, which are already updated to propagate the freshness.

In this stack I evaluated parts of our codebase **in bulk**, then I either changed the Freshness arg or removed the TODO comment.

## This diff
These are actually 2 callsites that I believe we could likely change to read from a replica...
Both are running all the time (thus could be contributing to unnecessary connections to the primary) and sound like should be fine with a latency of up to 1s (P95 is ~600ms).

RFC:

- RajivTS for repo_stats_logger
- markbt and YousefSalama for walker

Differential Revision: D81447235

fbshipit-source-id: 007d2300d5eec0faa07bff7ea05663e6ed4036cf