mysql: Read bookmarks from replica in modern_sync
Summary:
## This stack
In D81238781 I changed the `get` method in the `Bookmarks` facet to take a `Freshness` arg.
This was needed to then change the behaviour of always reading from the primary in D81212709.

I wanted to make the change no-op, so I updated all the callsites adding `Freshness::Recent`, which I would then later re-evaluate.

The problem is that there are **a lot of callsites**. Too many to evaluate one-by-one.
Besides that, most of the problematic calls are coming through SCS, via `repo_resolve_bookmark` or by calls to `mononoke_api::resolve_bookmark`, which are already updated to propagate the freshness.

In this stack I evaluated parts of our codebase **in bulk**, then I either changed the Freshness arg or removed the TODO comment.

## This diff

## RFC: modern_sync

~~Modern sync seems to follow bookmark update log ids, so worth reading from primary.~~
lmvasquezg and andreacampi to confirm.

**EDIT:** we'll actually read from the replicas. See discussion in comments.

Please keep in mind, like I mentioned in D81447238, that **if you believe you can get bookmark values from a replica** (e.g. if you think it's worth a try), **it can significantly reduce latency of some operations**, as I am seeing in [an ongoing experiment](https://fb.workplace.com/groups/3112141815538647/permalink/24488073400852179/), where the impact on latency is big for some regions: https://fburl.com/canvas/8mwajzly

Differential Revision: D81448031

fbshipit-source-id: 9e067c0cdfc979bc6a198273875614450da0b231