git_source_of_truth: Add a reserved state
Summary:
Since we will use the `git_repositories_source_of_truth` table to manage assigning repo ids in `create_repos` later down this stack,
we will need a state that represents a repository being "reserved", which the repo will be in while it is being created in mononoke.

Within `create_repos`, we will transition from this state to `mononoke` if the repo was successfully created.

Reviewed By: chiara125

Differential Revision: D81599089

fbshipit-source-id: a42d4df35b7e9fab51f41b804e4e71f2906d5938