git_source_of_truth: rename `set` to `insert_or_update_repo`
Summary:
`set` is quite vague, and I need it to be more precise so I can add a similar method and express the distinction.

* The behaviour is to insert a row or replace a pre-existing row, hence the `insert_or_update` part of the name.
* It applies to a single repo, hence the `_repo` part of the name.

The next diff in this stack introduces `insert_repos`, which insert multiple repos and fails if the id or name which are unique keys already exist.
Having renamed this beforehand makes the naming consistent and precise.

Reviewed By: clara-9

Differential Revision: D81598219

fbshipit-source-id: f4395417b6dfd8cb5abd8d1c7a56c6ff27cf05d0