store: make PidFetchCounts scalable
Summary:
Every ObjectStore access that misses the cache will attempt to update the fetch
counts in the PidFetchCounts structure, which requires taking a write lock and
then incrementing a counter. However, this means that on highly concurrent jobs,
like on a Watchman subscription when SHA1 are requested and the data isn't in
the in-memory caches, the write lock is heavily contended, leading to Eden
spending a large amount of time waiting on the lock.

Taking a step back however, the write lock is only really is needed when
inserting the first access for a given PID, as the insertion is the operation
that may rebalance/shuffle the map. For every other subsequent access, only the
count needs to be protected against concurrent writes, but that can be achieved
with an atomic. This latter case is by far the most common as applications are
likely to access the repo more than once.

Thus the strategy is to simply take an uncontended read lock, and increment the
count then, and if the PID wasn't present, then take the write lock. This should
virtually eliminate the lock contention.

Reviewed By: genevievehelsel

Differential Revision: D81596222

fbshipit-source-id: 8352f2384438eab1c84e24cd8fcacbcb2a659104