move heartbeat logic into its own class
Summary: I am planning to add more logic to the silent exit logger. Currently, the logic of this logger is distributed between `EdenMain` and `EdenServer` and makes it hard to read the code for future development. In this diff I want to move the heartbeat logic to its own class. Then any future development will in in one place.

Differential Revision: D80567098

fbshipit-source-id: fd0dcebb8a9dc8c5777fa5cbfd3aa6fb032824a2