Use sync feature flag to fix test
Summary: Merge conflict checking was using an async atom to check a feature flag. This suspends, which is not really expected in the conflicts UI. That breaks our test. This code path does not depend on it being async, and it's better to just use the sync version where it's false until the flag is checked.

Reviewed By: nsblake

Differential Revision: D81705993

fbshipit-source-id: 9dffab513448a4abc3c2b359c5db8ad69b646464