create_repos: Implement the contention resolution logic
Summary:
Set-up the structure of `create_repos` so we use the `git_repositories_source_of_truth` table as a
source of truth for which repo ids are available, and allowing us to reserve repo ids to avoid contention
during repo creation.

We still need to implement the actual creation of a repo config for these repos in configerator.

That will come further down the stack.

Reviewed By: RajivTS

Differential Revision: D81784998

fbshipit-source-id: 2723f2b9435f2cd1f46f763774b2ce4213376de0