Add recommended bookmarks section to Bookmarks Manager (V1)
Summary: The Bookmarks Manager lets users manually toggle which bookmarks appear in the main view. We want to provide a way to show the to show a “recommended” bookmark(s) that has higher cache hit rates for build and stability. For now, we pick a static bookmark (remote/fbcode/warm) to show to a high-priority cohort (AI). In a future diff, we look to retrieve this recommended bookmark based on data like build history, codehub project, etc.

Reviewed By: nsblake

Differential Revision: D81634918

fbshipit-source-id: 0af99345e91f7e607278a499e1b197805b2e45e5