store: shard blob/tree aux data caches
Summary:
The blob and tree in memory aux data caches are heavily contended due to being
behind a single lock. This can be seen when creating watchman subscription that
requires SHA1 for all file. When that happens, multiple Eden threads will try to
read and write to these caches and contend on it, leading to most of Eden's time
being spent waiting on the lock rather than actually doing any useful work.

Reviewed By: MichaelCuevas

Differential Revision: D81596221

fbshipit-source-id: e697150b06df170bff1bb4e32266f109e5b886fd