integration: refactor filteredhg tests
Summary:
# This diff

Refactors the FilteredHg tests so that a base FilteredHg class is created, and a FilteredFSBasic class is derived from the new base class. This is needed since the next diff introduces new flavors of the FilteredFSBase class

# Context:

It's believed that the cached Repo objects that FilteredFS uses may be holding onto indexedlog references and therefore keeping indexedlog files open. This would cause mmap'd indexedlog files to never be freed from memory, and therefore use a large amount of file-backed memory in perpetuity. Let's guard against this by evicting Repos from the cache after a configured amount of time.

Reviewed By: muirdm

Differential Revision: D81743177

fbshipit-source-id: e092dfbaa066e4a43e3eec1cc011526d49a1955b