Utilise Vec<u8> instead of Bytes in Mononoke Git for transferring data
Summary:
Usually, its a good choice to use `Bytes` when dealing with byte level data (i.e. slice of u8) because if you accidentally clone it, you just clone the reference and not the underlying data. Additionally, you don't have to worry about lifetime issues because its essentially similar to a `Arc<Vec<u8>>` under the covers.

However, in Mononoke Git read path its actually a source of problems because eventually we want to send the bytes to the client. Using bytes as an intermediary results in multiple layers of accidental copies (e.g. when converting from PackfileItem to Entry object type expected by gix or when thrift deserializing BasePackfileItem from blobstore bytes). This diff addresses that problem

Reviewed By: YousefSalama

Differential Revision: D81787640

fbshipit-source-id: 32247933b130a9c81f7bea4c7501d2922983b6ac