Implement support for memory based buffering in Mononoke Git
Summary:
Stream buffering based on number of futures is an approach that we use everywhere in Mononoke. However, Mononoke Git is a special snowflake where the blobs that we pull from the blobstore can range anywhere from 100 bytes to 10 GB in size. This makes it very very tricky to figure out a number that gives good throughput while ensuring we don't buffer too much.

This diff ensures that we start buffering based on the resource that matters, i.e. the memory usage.

Reviewed By: YousefSalama

Differential Revision: D81787644

fbshipit-source-id: 15fba33ef8af63925ea676a0c7107b2321f75a93