Allowlist WEKA as a valid filesystem to mount Eden onto
Summary:
See https://fb.workplace.com/groups/778748051005055/posts/24377683465188272 for more discussion.

-----

We have deployed the WEKA filesystem in the Corp DCs that host the Silicon Development environment.  As part of using this to replace existing NFS-based workflows, we would like to start experimenting with hosting Hargow workspaces on these WEKA filesystems, which entails needing to be able to mount Eden repositories on top of this filesystem.

NOTE: We're not expecting Eden to natively synchronize across multiple hosts across the WEKA mounts, as with our existing NFS use cases, we expect to be using Hargow's synchronization logic to accomplish this.  All that we need is for Eden to support mounting on top of this filesystem.

This diff enables Eden to mount on top of WEKA filesystems, which just entails allowlisting the WEKA filesystem type.

Reviewed By: MichaelCuevas

Differential Revision: D81365146

fbshipit-source-id: 6bd1deec853899a66a42add078b17e0026df9281