sql_ext: Read bookmarks values from replica in cache_warmup::do_cache_warmup
Summary:
See [this post](https://fb.workplace.com/groups/1206130147756479/posts/1322679599434866).

**TL;DR**: reading from replicas is likely to bring **significant performance improvements**.

This doesn't seem to be latency sensitive, e.g. require RAW. TBF, it also likely doesn't contribute much to a high number of connections, since it's like done when starting up a task, but if we can read from a replica, why not?

Maybe YousefSalama or clara-9 know if we could do this? you're the two people that, IIRC, have worked with microwave or mutable storage 😅

Reviewed By: andreacampi

Differential Revision: D81790219

fbshipit-source-id: 3734e54cacd792dd7ac8457856caa1ae9a9b38b3