dirsync: demostrate rebase issue with custom mirrored changes
Summary:
If a commit contains changes that are mirroed with a custom script, it can
error out rebasing (with in-memory disabled). The added test demostrates the
issue.

Note: In-memory rebase has a bug that causes dirsync to not able to load the
`.hgdirsync` config file from its `memctx`. Therefore this issue might not always
happen in production.

Reviewed By: muirdm

Differential Revision: D82039560

fbshipit-source-id: d5010951462dfb0a0b0148765cc5fa3504ee96a6