Add visibility on recommended bookmarks when opting in to use them
Summary: In the Bookmark manager, it would be helpful to the user to see which bookmarks are being recommended when the user clicks the checkbox. This diff highlights the recommended bookmarks and automatically checks them (disabling the others) when the user decides to use them via the checkbox. Additionally, the recommended commits are prioritized in the Bookmarks Manager. The next diff adds a tag to distinguish bookmarks as recommended.

Reviewed By: nsblake

Differential Revision: D81926265

fbshipit-source-id: 04a3bef030acf0f118892581ce1e6c39ec6293f1