sql_ext: Standardize experiment JK logging
Summary:
This generalizes what was done in D76728908 and D81212709.

Basically it simplifies adding more JKs for running experiments, along with their switches.

This is definitely not the most beautiful thing, but it beats creating a column for every new JK I create...

I'll be running multiple experiments for XDB performance, so this will definitely be helpful.

Reviewed By: RajivTS

Differential Revision: D82029910

fbshipit-source-id: 4fd66e8135b6c8835783507fd4cea69680bd8ca1