Add JK to disable read fallbacks to primary replica
Summary:
See T237440708 for more context.
RajivTS and I have been trying to understand if this fallback is actually needed...

In D82029194 I added ODS counters for us to confirm if these fallbacks are actually needed.

If they aren't, we'll run try disabling the fallback to measure the impact on latency.

Reviewed By: RajivTS

Differential Revision: D82031634

fbshipit-source-id: 17fe153288d2193535b827b57ee53980a0d69f8a