Add recommended bookmark icon for user information
Summary: In opt-out mode, where users will be automatically opted in to use the recommended bookmark, we want to give some information on why only a few bookmarks are shown. This adds a recommended icon at the bookmark level in the commit tree and in the remote bookmarks list with a tooltip to give users information about the recommended bookmark and how they can manage them.

Reviewed By: nsblake

Differential Revision: D81817106

fbshipit-source-id: 7ddcd26fbed97b6c9498f16d8226b2ac8357741a