git_delta_manifest_v3: allow fetching manifest blob directly without having to fetch mapping first
Summary: We don't benefit much from having the GDMV3 manifest blobs be content addressed as they only store deltas corresponding to the changes a commit made. Change the key for the GDMV3 blobs to use the changeset ID instead of the hash of the content, and allow fetching the blob directly instead of having to fetch the mapping blob first. This should give a speed up due to not having to make a second blobstore round trip to be able to fetch each GDMV3 manifest blob.

Reviewed By: RajivTS

Differential Revision: D82016080

fbshipit-source-id: ceef032349b25ffd8ba81821be6d185d6d138adc