Remind users to use recommended bookmarks
Summary: If users un-check the recommended bookmark, they may forget to check it again. If we detect this, we show a message after some time (2 days) prompting users to check the box again. This is done via a manual tooltip that checks if the user has enabled recommended bookmarks and when they were last shown the prompt

Reviewed By: nsblake

Differential Revision: D82139904

fbshipit-source-id: 67fa3804ec3799cdfd8aafc91529aea0962d52fb