AI Code Review Comments
Summary:
This is a rough attempt to plumb AI "first-pass" comments through the existing ISL inline comments architecture. The architecture mainly consists of a few core pieces:
- `InlineCommentsProvider`: Main class that provides inline comments
- `InlineCommentsController`: Manages comments data and fetching
- `DiffCommentPanelProvider`: Manages how comments appear in the sidebar panel
- `InlineCommentsInsetProvider`: Controls how comments appear inline in webview insets

The general idea behind this diff was to re-use as much existing architecture as possible while minimizing re-factoring / duplication. In this case, the main thing that needs to change is the data layer and how comments are fetched (since they no longer can be queried from Phabricator through GraphQL, so I decided to extend the `InlineCommentsController` with a first-pass-specific implementation that changes the bare minimum functions necessary to render comments in the required spaces.

Data management is handled via a `localStorageBackedAtomFamily`, which in VS Code will read/write from `globalStorage`. I think it makes sense to have this so that we can manage client-level state per comment easily (e.g. to show the number of comments / previews for a given commit), and then on the extension side we just need to ensure that we use the same keys when updating `globalStorage` from other sources (e.g. from DVSC through the API).

Message-passing was a bit weird - I ended up using the normal client-server message passing to pass the review comments from the ISL client entrypoint (where I'm currently running a crude review) to the server, and then I pass those results to the controller via a VS Code command. This will definitely need to be cleaned up to work with the planned API.

Reviewed By: evangrayk

Differential Revision: D82125961

fbshipit-source-id: ccee1c250efef60625e79f799db7f70a2166189f