storemodel: change TreeEntry::iter() to yield borrowed data
Summary:
Change TreeEntry::iter() to yield `&PathComponent` instead of `PathComponentBuf` to avoid forced allocation iterating tree entries. This is an optimization for tree processing in eden where the owned data is not helpful since it needs to be copied again anyway.

I left the allocating iter around as `iter_owned()` since some py bindings rely on getting `'static` data when converting the iterator to a py iterator.

Reviewed By: zzl0

Differential Revision: D81650019

fbshipit-source-id: e17d4fdb7a882684fbcb146cf3cad51f826875e0