backingstore: pipe through object id format and case sensitivity
Summary:
These two tidbits are required when creating trees (which I am going to move further down the backingstore). Pipe them through down to the SaplingNativeBackingStore. I added an ffi enum for HgObjectIdFormat, but not for CaseSensitivity because it seems like bool repr enums aren't supported in Rust.

Note that SaplingBackingStore.cpp currently uses kPathMapDefaultCaseSensitive when constructing trees, which doesn't seem correct. It should be using the case sensitivity value from the CheckoutConfig.

Differential Revision: D81650016

fbshipit-source-id: e39b65e615dc1e9131ed01b04005ead916f5d41d