metrics: delay sync to ODS
Summary:
Bumping the ODS counters (and especially bumping the OBC stuff) uses a lot of CPU. Smooth things out by instead updating ODS counters 0.1% of the time, and otherwise updating ODS every 60s when eden flushes the backingstore, and via the flushStatsNow() thrift call.

I implemented it by adding a new atomic counter to store the last value we synced to ODS. This way the "hot" path only has a single atomic increment, but the downside is we need a compare-and-swap loop when updating ODS.

Even for just the plain ODS counters we spend 30% of our ScmStore get_local_aux_direct time CPU time incrementing ODS counters. With this change, the counter increment doesn't show up in the CPU profile anymore.

Reviewed By: MichaelCuevas

Differential Revision: D81721567

fbshipit-source-id: 9545c3f7d6eb5f891cafa51af90f8f623daf9488