manifest-tree: add TreeEntry::size_hint() trait method
Summary:
This returns the number of tree entries, if available.

I'm going to use this to optimize tree entry allocations as eden processes sapling trees.

The interesting hint is for the "basic" hg tree. We use the SIMD bytecount crate to quickly count the entries using the "\n" record terminator.

Reviewed By: zzl0

Differential Revision: D81650013

fbshipit-source-id: bc143ebe42bbfcf2ff70df2a34734ede989ccb44