backingstore: use new TreeBuilder when fetching single trees
Summary:
The SaplingNativeBackingStore creates a new TreeBuilder object and passes that to Rust, who populates the Tree directly as it iterates over the Rust `TreeEntry` trait object. This skips creating the ffi::Tree object, and is more efficient than the previous fromRawTree() conversion in SaplingBackingStore.cpp.

Also, we use the correct case sensitivity (from CheckoutConfig) when building the Tree instead of assuming kPathMapDefaultCaseSensitive like SaplingBackingStore.cpp.

Reviewed By: MichaelCuevas

Differential Revision: D81650012

fbshipit-source-id: c854153a1cb48b75dfbaa5f9fb025ab3e00eea5b