vfs: accept Blob for write instead of &[u8]
Summary:
Change the VFS to accept our blob::Blob type instead of &[u8]. This allows it to operate more efficiently on LFS files, which are represented as a chain of mmap buffers via IOBuf. Previously we would copy the entire LFS content into a `Vec<u8>` so we could write it out, but now we can write out the IOBuf chunks directly.

The main user of VFS is checkout, but that code already had a Blob in-hand, so there wasn't actually much stuff to change.

Note that there was some weirdness with propagating errors from flume::Sender. The error contains the item itself ("Blob" in this case), which is not `Send`, so I had to tweak things.

Reviewed By: zzl0

Differential Revision: D82282021

fbshipit-source-id: ed2a42f62f38ff47dddf5245807dd4a3cba2c3f4