blob: pull out "each_chunk" iterator fn
Summary:
Pull out a callback-based iterator of the blob's chunks. The IOBuf can be made up of non-contiguous chunks, and this allows iterating them all without copying.

I made a minor optimization to avoid the calls to `cur.has_remaining()` by assuming the iterator is done iff the returned chunk is empty. This is the documented behavior of bytes::Buf (which our IOBufShared Rust object respects).

I tried to make a real Iterator, but I couldn't get the lifetimes compiling.

Reviewed By: zzl0

Differential Revision: D82282016

fbshipit-source-id: 575f59fb91b0f39a5b4cd1eba55e8db0d05a7754