blobstore: get rid of BlobstoreUnlinkOps trait and make unlink method part of the Blobstore trait
Summary:
`BlobstoreUnlinkOps` is meant to be a "mixin" kind of trait for blobstores that additionally support unlinking a key, but in reality all our blobstores support the unlink operation so this trait only adds additional boilerplate. This diff removes it and adds the unlink method to the `Blobstore` trait instead.

The only kind of blobstore that doesn't support unlink is the `MemWritesBlobstore` that writes to memory instead of writing to the underlying storage. I think any semantics we choose for unlink for it are going to be unintuitive so I opted for returning an error instead.

Reviewed By: lmvasquezg

Differential Revision: D82436290

fbshipit-source-id: a9cb82dc41da311bc7361f0dc8b7eccb7ddeb867