Make recommended bookmarks the default
Summary:
To encourage adoption, recommended bookmarks will be a default and we will have this be an opt-out model. It will be gated by GK.

This adds some logic to the implementation, as we can no longer rely on useRecommendedBookmark atom as a proxy for the GK (previously the only way to modify useRecommendedBookmark atom was through a GKed component, and the atom was defaulted to false).

Reviewed By: nsblake

Differential Revision: D82246898

fbshipit-source-id: 60459d6eaab9134f48ead0b58de226b1a25122f2