sql_ext: Split query method to get access to telemetry
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

The first application I intend to use this with is T234601360, a SEV follow up about disabling these fallbacks to primary for filenodes reads.

## This diff (no-op)
Split the query method so that the new method I introduce in D82226564 to retry the read to replicas is able to access the response attributes.

Differential Revision: D82285997

fbshipit-source-id: 37289e9a2c680ac3ffd7634a820366e009638872