sql_ext: Extract error counter bump into function
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

The first application I intend to use this with is T234601360, a SEV follow up about disabling these fallbacks to primary for filenodes reads.

## This diff (no-op)
Extract snippet that bumps error counters in ODS to separate function, so we can re-use it in D82226564 and avoid duplication.

Differential Revision: D82426887

fbshipit-source-id: 0bc01934aefc877e73029ddf476efe8ae58008c3