Pass --user when amending
Summary:
If you download someone else's draft commit and amend it, we can mark it as being your commit now.

This will update the commit's icon to be your avatar, which makes it a little more clear you've made changes.

We only pass this arg if your username differs from what is in the commit. This makes the cli arg preview a little cleaner. It's easy to do because we need to have the username to pass in the first place.

Reviewed By: alichtman

Differential Revision: D82463157

fbshipit-source-id: c5b2d4dbe444592da59f86bf1d4b6b1994c79b3f