Add context menu option to copy selected diff IDs to clipboard
Summary:
When > 1 diffs are selected, show a new context menu option that allows the user to copy the IDs of the selected diffs to clipboard.

I find myself needing this feature relatively often and have found it pretty tedious to copy the IDs one-by-one.

I'm not aware of any similar solutions to this problem.

Reviewed By: evangrayk

Differential Revision: D82471559

fbshipit-source-id: 7c0e3c40a97ef5eb8b74ea0e744c1fe05292d091