filteredfs: fix interrupted checkout message
Summary:
Fix the interrupted checkout advice message to not display the RootId directly, instead asking the backing store to render it.

The SaplingBackingStore returns a 20 byte binary commit hash, so we need to hexify it. It is somewhat surprising that "rendering" the RootId (which is a hex string) produces a binary hash.

Reviewed By: MichaelCuevas

Differential Revision: D82550458

fbshipit-source-id: 5f775d0f40cd9b7e6d35f9913b3959cfea40de43