filenodes: Update unit tests to use replication lag instead of manually deleting data from replica
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

The first application I intend to use this with is T234601360, a SEV follow up about disabling these fallbacks to primary for filenodes reads.

## This diff

In D82431553 I tried to use the new `query_with_consistency` method, but a couple of tests failed.
These two tests failed because they test the logic of falling back to the primary for reads if the data is not found in the replica.

**However**, they do that by manually deleting the data from the replica.

In this diff I change both tests to use an artificial replication lag. This will ensure that the fallback to master always happens, as you'll be able to confirm with the new `QueryCounters` that I defined.

These counters implement `SqliteCallbacks` and are incremented after every read or write query.

In D82431553, you'll see how the counter values change, because we stop falling back to the primary for reads.

Reviewed By: chiara125

Differential Revision: D82437980

fbshipit-source-id: dd70142838e26fbb485c8eac508e54c683051821