sql_ext: Introduce query_with_consistency method
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

The first application I intend to use this with is T234601360, a SEV follow up about disabling these fallbacks to primary for filenodes reads.

## This diff
Introduce a new query method that will leverage the HLC returned in the response attributes (see [this MySQL wiki](https://www.internalfb.com/wiki/Infra_Cloud/Databases/MySQL_XDB/MySQL_Clients/Configuring_Read_your_own_Writes/)) to check if the replica was up-to-date when it served the request.

If it wasn't, the query would be retried up to a certain amount of times.

## Notes
- I confirmed with MySQL team that the `hlc_ts_lower_bound` returned in the response attributes represent the last update from the replica.
- I'm talking to them about another option that would be even better, but I don't believe that it's possible right now: manually setting the `hlc_ts_lower_bound` in the **query** attributes and using a DbLocator with `InstanceRequirement::ReadAfterWriteConsistency` to get the same behaviour as read your own writes, **without writing anything**.
  - This would allow us to also configure how much a replica would wait for replication to catch up.
- I decided to implement the first option right away because we would need to introduce the API **either way**. If we're able to try the other option, the code to write would be minimal and we would only change the implementation of `query_with_consistency` without affecting the existing callsites.
- I only implemented this new method for the read and **no cache** variant, since the callsites that will use it are looking for very fresh data, so the queries are not cached.

Differential Revision: D82226564

fbshipit-source-id: 6f97972d7c39c4e9b7e0938ad834d25bb9f050a3