sql_ext: Helper function to determine if query_with_consistency should be use and provide params
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

The first application I intend to use this with is T234601360, a SEV follow up about disabling these fallbacks to primary for filenodes reads.

## This diff
We'll be updating many callsites to use `query_with_consistency`. Each callsite will need to read multiple JKs to check if the feature is enabled and determine the parameters for retries, etc.

I decided to introduce a helper method right away, which will reduce code duplication.

The function will return `None` if the callsite should not use `query_with_consistency`.

Reviewed By: chiara125

Differential Revision: D82432897

fbshipit-source-id: 9cbe872162bfa4f7afebb0c3ddb0d72962aa45f7