edenapi: use final reponses headers after redirection
Summary:
There's a bug when trying to use Commit Cloud Git on corp, where we'd get a deserialization error. For context, to reach the Mononoke Git server on Commit Cloud Git, we first go through the location service, which returns a 307 (temporary redirection), and we follow that redirect to get the end result.

 After some debugging, the bug turned out to be that, given that we use `zstd` encoding in corp, the server was returning the appropriate `zstd` envelope and headers, but the client was only reading the headers from the initial redirection, hence trying to decode the response with a different encoding.

 This diff basically resets the response-decoding state between hops, making sure that if we have a redirection, we read the final headers and body.

Reviewed By: muirdm

Differential Revision: D82447141

fbshipit-source-id: c879803f578fe6e54cddd2e405742ade98a7ea66