sql_ext: Stop using --mysql-master-only by default on integration tests
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

The first application I intend to use this with is T234601360, a SEV follow up about disabling these fallbacks to primary for filenodes reads.

## This diff

When I changed the behaviour of filenodes fetches in D82431553, multiple tests started failing.
The unit tests I addressed in D82437980.
This diff addresses the issue with the integration tests.

All the integration test that run with MySQL are using the `--mysql-master-only` flag, which uses the master connection for reads.
I'm changing this default behaviour, because **using replicas for reads better test the behaviour in production**.

This change will also work properly with the changes from D82431553.

Reviewed By: andreacampi

Differential Revision: D82543678

fbshipit-source-id: b7d2ab9af98aad8e104a377f98209274a167f004