Add more information for accessing bookmark manager on recommended bookmark tooltip
Summary:
Since recommended bookmarks are opt out, we want to give more information when hovering on the new recommended bookmark on how to toggle the bookmarks manager. This diff adds information for the shortcut users can use.

It is also possible if many users are confused, we can have the bookmark be clickable. However this ends up being a bit hacky because of the tooltip wrapper which seems to interefere with click events. So we want to try this first in dogfooding.

Reviewed By: LeVuHuy

Differential Revision: D82647219

fbshipit-source-id: a86288edd3e2ee2182d9bb82a91ba6ef0319ad2e