mysql: Log how often primary read is useful for bonsai_hg_mapping
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

## This diff

Like I did for git mappings in D82029194, let's track how often we're actually finding something in the primary that was not found in the replica, i.e. how often we don't find something because of replication lag.

Differential Revision: D82548193

fbshipit-source-id: c67357cd8bb151c0a41d4e2b0f80dc8c059f9672