sql_ext: Fallback to primary for reads if HLC is missing
Summary:
## This stack

This aims to address a common pattern across our codebase: falling back to the primary replica for reads where you don't find the data you're looking for.

The following tasks have a lot of details: T237806453, T237287313.

The first application I intend to use this with is T234601360, a SEV follow up about disabling these fallbacks to primary for filenodes reads.

## This diff

In D82226564 I introduced the `query_with_consistency` method. Before shipping, I made a change (V4) where, if the HLC was not returned in the response attributes, as its expected **in MySQL** responses, the method would throw a `ConsistentReadError::MissingHLC` exception, instead of just returning the query result it got.

This is the right approach, as any calls to `query_with_consistency` **should assume that if it returns any result, the result is up-to-date** and we can only do that with the HLC.

However, in practice, it broke a ton of tests, because Sqlite doesn't return the HLC unless we mock it in the tests (D82437980).

## Solution
Change `query_with_consistency` to take a reference to `SqlConnections`, which contains both read and read_master connections.
If the query fails with `MissingHLC`, we fallback to reading from the primary, **as its currently done in all callsites that we'll be updating to use `query_with_consistency`**.

In this process, we'll also bump ODS counters and log something to ensure that we know if/when/where this is happening.

Differential Revision: D82611356

fbshipit-source-id: 4d1666387c12c50e924772043c97b5911f8ab10a