InferredCopyFrom: add a cap on number of changed files to consider
Summary:
Since inferred_copy_from complexity can be `O(N*N*M)` where N is number of changed files, M is file size, we want to put a cap on N.
The previous diffs have made this into the derived data config. This diff is applying it to the logic.

To make things easy to reason about, the cap is directly applied after `bonsai.simplified_file_changes` for all scenarios.

Reviewed By: RajivTS

Differential Revision: D82635504

fbshipit-source-id: 0a863cae28f1e4b981ecf3b4563ddd2846a88b9d