FilteredBackingStore: optimize unfiltered case
Summary:
Optimize FilteredBackingStore to use the "raw" underlying BackingStore's ObjectIds when possible. This avoids lots of copies and allocations. It is like a "specialization" when the underlying backingstore is a SaplingBackingStore (which is basically "always").

FilteredBackingStore has a major perf impact fetching trees since it deep copies every tree, allocating ObjectIds and performing `O(log(# entries))` work for every TreeEntry. The vast majority of Trees are unfiltered, though, so this is very wasteful.

The main reason it copied Trees was so it could change the Tree's entries' ObjectIds to the FilteredObjectId format. Basically, it adds a "type" byte to the beginning of the underlying ObjectId.

To avoid all the copying, we can simply re-use the underlying BackingStore's ObjectId for unfiltered trees. The only issue is we need to be able to tell the underlying ObjectIds apart from FilteredObjectIds. Fortunately, FilteredObjectId already uses a non-overlapping set of leading "type" bytes than HgProxyHash, so we can tell them apart by simply looking at the first byte. This only works, obviously, if the underlying BackingStore is a SaplingBackingStore. Perhaps we should formalize the ObjectId-first-byte-identifies-BackingStore rule so we can tell ObjectIds apart more confidently.

This optimization results in 2x faster "eden glob" performance, almost on par with non-filteredfs.

The new optimization is gated by the experimental.filteredfs-optimize-unfiltered flag (defaults to true in tests, false otherwise). When enabled, the flag causes unfiltered trees to be returned directly by FilteredBackingStore, "activating" the optimization. Even with the flag disabled, the FilteredBackingStore still supports passing through underlying SaplingBackingStore ObjectIds. This is in case something goes wrong and we have to disable the optimization - we don't want the SaplingBackingStore ObjectIds to suddenly start causing errors.

Reviewed By: MichaelCuevas

Differential Revision: D82507321

fbshipit-source-id: de3cc8770c59577b89dc7c1d02db88849912f7f2