scs: add commit_directory_branch_clusters method
Summary:
Add a new method to SCS that will return the list of directory branch clusters in a commit.  A directory branch cluster is a group of directories that are directory branches of each other.  One directory is deemed to be the main, or primary, directory of the cluster.  This can be used to favour a particular branch when deduplicating branched files.

The current implementation uses a fixed configuration of clusters, filtering out paths that don't exist in the repo.  In the future we will add a derived data type that will auto-detect directory branches.

Reviewed By: gustavoavena

Differential Revision: D82411359

fbshipit-source-id: 4b839f229c031fdde0bb9d754de0653f45da4d32