slapi: populate "is binary" field in LFS pointers
Summary:
Sapling uses this to cheaply know if an LFS file contains binary data. When this metadata is not present, Sapling assumes is_binary=true, which results in annoying behavior for text files (such as "sl blame" erroring out without the "-a" flag).

Sapling has always supported parsing the "x-is-binary" LFS pointer flag, so this change should be relatively safe.

Reviewed By: markbt

Differential Revision: D82591323

fbshipit-source-id: 1cc8441e333c881f34c2dd48d42b2de660dc9bca